pyproject.toml
[project]
name = "teste"
version = "0.1.0"
description = ""
authors = [
    {name = "bezerraluiz-codeform",email = "luiz.a@cartoriocriciuma.com.br"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "langchain-text-splitters (>=1.1.0,<2.0.0)",
    "langchain-huggingface (>=1.2.0,<2.0.0)",
    "sentence-transformers (>=5.2.0,<6.0.0)",
    "langgraph (>=1.0.5,<2.0.0)",
    "langchain-openai (>=1.1.3,<2.0.0)",
    "langchain-core (>=1.2.1,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "tavily-python (>=0.7.15,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
